refactor: simplify get_token_mapping
The `get_token_type_mapping` function is used to get different token types for public repos.
These depend on install YAML configuration. It only affects public repos that don't have a token defined(*)

So the function used to make 2 checks:
* repo is private?
* github app installations?

With the refactoring from the previous commit fetching the github app installation info was pulled to `get_repo_provider_service` function.
So we can decide there if we need the token type mapping for this case or not.
This simplified the `get_token_type_mapping` function, which is preferable.

I also removed the unused `commit` arg from `get_repo_provider_service`.

(*) - except for comments, which prefer the configured commenter bot. I don't know why
