+++ tasks/commit_update.py
@@ -42,7 +42,7 @@ class CommitUpdateTask(BaseCodecovTask, name=commit_update_task_name):
                 db_session, self.name, repository.owner
             )
             repository_service = get_repo_provider_service(
-                repository, commit, installation_name_to_use=installation_name_to_use
+                repository, installation_name_to_use=installation_name_to_use
             )
             was_updated = async_to_sync(possibly_update_commit_from_provider_info)(
                 commit, repository_service
